+++ sdk/communication/Azure.Communication.JobRouter/src/Models/ExceptionPolicy.cs
@@ -27,6 +27,22 @@ namespace Azure.Communication.JobRouter
         /// <summary> (Optional) The name of the exception policy. </summary>
         public string Name { get; set; }
 
+        [CodeGenMember("Etag")]
+        internal string _etag
+        {
+            get
+            {
+                return ETag.ToString();
+            }
+            set
+            {
+                ETag = new ETag(value);
+            }
+        }
+
+        /// <summary> Concurrency Token. </summary>
+        public ETag ETag { get; internal set; }
+
         void IUtf8JsonSerializable.Write(Utf8JsonWriter writer)
         {
             writer.WriteStartObject();
@@ -44,6 +60,11 @@ namespace Azure.Communication.JobRouter
                     writer.WriteObjectValue(item);
                 }
                 writer.WriteEndArray();
+            }
+            if (Optional.IsDefined(ETag))
+            {
+                writer.WritePropertyName("etag"u8);
+                writer.WriteStringValue(ETag.ToString());
             }
             writer.WriteEndObject();
         }